requirements.txt

finnhub-python
pandas


websocket-client

